# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: SumUp-inspired payments platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /onboarding/session:
    get:
      operationId: getOnboardingSession
      summary: Get current onboarding session
      tags: [onboarding]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Session state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
    post:
      operationId: createOnboardingSession
      summary: Start a new onboarding session
      tags: [onboarding]
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
  /onboarding/session/step:
    post:
      operationId: submitOnboardingStep
      summary: Submit an onboarding step
      tags: [onboarding]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingStepRequest"
      responses:
        "200":
          description: Step submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingSession"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /onboarding/session/kyc/verify:
    post:
      operationId: submitKycVerification
      summary: Submit KYC documents for verification
      tags: [onboarding]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitRequest"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycVerificationResult"
...
  /onboarding/question-tree:
    get:
      operationId: getQuestionTree
      summary: Get onboarding question tree
      tags: [onboarding]
      parameters:
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: business_type
          schema:
            type: string
      responses:
        "200":
          description: Question tree
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuestionTree"

  # ─── TRANSACTIONS ──────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
        message:
          type: string

    # ─── AUTH ────────────────────────────────────────────────────────────────
...
    OnboardingSession:
      type: object
      required: [id, merchant_id, current_step, status, answers, created_at]
      properties:
        id:
          type: string
          format: uuid
        merchant_id:
          type: string
          format: uuid
        current_step:
          type: string
        status:
          type: string
          enum: [not_started, in_progress, kyc_pending, kyc_approved, banking_complete, submitted, approved, rejected, needs_more_info, manual_review]
        answers:
          type: object
          additionalProperties: true
        next_question:
          $ref: "#/components/schemas/OnboardingQuestion"
        progress_percent:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    OnboardingQuestion:
      type: object
      required: [id, type, question]
      properties:
        id:
          type: string
        type:
          type: string
        question:
          type: string
        hint:
          type: string
        field:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuestionOption"
    QuestionOption:
      type: object
      required: [value, label]
      properties:
        value:
          type: string
        label:
          type: string
    OnboardingStepRequest:
      type: object
      required: [step_id, answer]
      properties:
        step_id:
          type: string
        answer:
          type: object
          additionalProperties: true
    KycSubmitRequest:
      type: object
      required: [id_type]
      properties:
        id_type:
          type: string
          enum: [passport, national_id, driving_licence]
        document_front_base64:
          type: string
        document_back_base64:
          type: string
        selfie_base64:
          type: string
    KycVerificationResult:
      type: object
      required: [status, session_id]
      properties:
        status:
          type: string
          enum: [pending, approved, rejected, manual_review]
        session_id:
          type: string
        message:
          type: string
...
    QuestionTree:
      type: object
      required: [version, root, nodes]
      properties:
        version:
          type: string
        root:
          type: string
        nodes:
          type: object
          additionalProperties: true

    # ─── PAYMENTS ────────────────────────────────────────────────────────────
...