# lib/api-zod/src/generated/types/kycVerificationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SumUp-inspired payments platform API
 * OpenAPI spec version: 1.0.0
 */
import type { KycVerificationResultStatus } from "./kycVerificationResultStatus";

export interface KycVerificationResult {
  status: KycVerificationResultStatus;
  session_id: string;
  message?: string;
}