# lib/api-zod/src/generated/types/onboardingSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SumUp-inspired payments platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OnboardingQuestion } from "./onboardingQuestion";
import type { OnboardingSessionAnswers } from "./onboardingSessionAnswers";
import type { OnboardingSessionStatus } from "./onboardingSessionStatus";

export interface OnboardingSession {
  id: string;
  merchant_id: string;
  current_step: string;
  status: OnboardingSessionStatus;
  answers: OnboardingSessionAnswers;
  next_question?: OnboardingQuestion;
  progress_percent?: number;
  created_at: Date;
  updated_at?: Date;
}