# lib/api-zod/src/generated/types/onboardingStepRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SumUp-inspired payments platform API
 * OpenAPI spec version: 1.0.0
 */
import type { OnboardingStepRequestAnswer } from "./onboardingStepRequestAnswer";

export interface OnboardingStepRequest {
  step_id: string;
  answer: OnboardingStepRequestAnswer;
}